post edit

+++ src/components/ui/PostCard/index.tsx
@@ -1,30 +1,35 @@
 import { FC, memo, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { PostsApi } from '../../../services/api/PostsApi';
+import { changeEditableId } from '../../../state/postEdit/slice';
 import { deletePost } from '../../../state/posts/slice';
 import { RootState, useAppDispatch } from '../../../state/store';
 import { Status } from '../../../types/fetchStatus';
 
 import LikeButton from '../LikeButton';
+import PostForm from '../PostForm';
 
 interface PostCardProps {
-  text: string;
   postId: string;
+  text: string;
   likesCount: number;
   postComments: any[];
   isLiked: boolean;
+  isOwnPost: boolean;
 }
 
 const PostCard: FC<PostCardProps> = ({
-  text,
   postId,
+  text,
   likesCount: likes,
   postComments,
   isLiked: liked,
+  isOwnPost,
 }) => {
   const dispatch = useAppDispatch();
 
   const { loggedIn } = useSelector((state: RootState) => state.auth);
+  const { editableId } = useSelector((state: RootState) => state.postEdit);
 
   const [likesCount, setLikesCount] = useState(likes);
   const [isLiked, setIsLiked] = useState(liked);
@@ -45,14 +50,30 @@ const PostCard: FC<PostCardProps> = ({
   };
 
   const handleDelete = () => {
-    // todo: add confirmation
+    const shouldDelete = window.confirm(
+      'Are you sure you want to delete this post?'
+    );
+
+    if (!shouldDelete) return;
+
     PostsApi.deletePost(postId).then(() => {
       dispatch(deletePost(postId));
     });
     // todo: add success notification
-
     // todo: add error notification
   };
+
+  const handleEdit = () => {
+    dispatch(changeEditableId(postId));
+  };
+
+  if (editableId && editableId === postId)
+    return (
+      <>
+        <PostForm type='edit' textareaValue={text} />
+        <button onClick={() => dispatch(changeEditableId(null))}>Cancel</button>
+      </>
+    );
 
   return (
     <div>
@@ -68,7 +89,12 @@ const PostCard: FC<PostCardProps> = ({
             isLiked={isLiked}
             handleLike={handleLike}
           />
-          <button onClick={handleDelete}>Delete</button>
+          {isOwnPost && (
+            <>
+              <button onClick={handleDelete}>Delete</button>
+              <button onClick={handleEdit}>Edit</button>
+            </>
+          )}
         </>
       )}
     </div>
